% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_datasets}
\alias{get_available_datasets}
\title{Get Available Eunomia Datasets}
\usage{
get_available_datasets()
}
\value{
A data frame with dataset names and descriptions.
}
\description{
Returns a list of available Eunomia datasets with descriptions.
}
\examples{
datasets <- get_available_datasets()
print(datasets)

}
