% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{plot_cdm_demographics}
\alias{plot_cdm_demographics}
\title{Plot CDM Demographics}
\usage{
plot_cdm_demographics(cdm, bins = 30, alpha = 0.7)
}
\arguments{
\item{cdm}{A CDM reference object created by \code{\link{setup_omop_cdm_eunomia}}
or a similar CDMConnector CDM object.}

\item{bins}{Integer specifying the number of bins for the age histogram.
Default is 30.}

\item{alpha}{Numeric value between 0 and 1 specifying the transparency
of the histogram bars. Default is 0.7.}
}
\value{
A ggplot2 object showing the age distribution by gender.
}
\description{
Creates demographic visualizations for an OMOP CDM database including
age and gender distributions.
}
\details{
The function creates a histogram showing the age distribution of patients
in the CDM, colored by gender. Ages are calculated assuming the current
year is 2024. Gender concept IDs are mapped to readable labels:
\itemize{
\item 8532 = Female
\item 8507 = Male
\item Other values = Other
}
}
\examples{
\dontrun{
# Setup CDM and create plot
cdm_setup <- setup_omop_cdm_eunomia()
plot <- plot_cdm_demographics(cdm_setup$cdm)
print(plot)

# Customize the plot
plot_custom <- plot_cdm_demographics(cdm_setup$cdm, bins = 20, alpha = 0.5)
print(plot_custom)

# Clean up
cdmDisconnect(cdm_setup$cdm)
}

}
\seealso{
\code{\link{explore_cdm}}, \code{\link{setup_omop_cdm_eunomia}}
}
