% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_eunomia.R
\name{setup_omop_cdm_eunomia}
\alias{setup_omop_cdm_eunomia}
\title{Setup OMOP CDM with Eunomia Datasets}
\usage{
setup_omop_cdm_eunomia(
  dataset_name = "GiBleed",
  cdm_version = "5.3",
  verbose = TRUE
)
}
\arguments{
\item{dataset_name}{Character string specifying the dataset to use.
Available options include:
\itemize{
\item "GiBleed" (default): Small dataset for testing
\item "synthea-allergies-10k": Synthea allergies data (10k patients)
\item "synthea-covid19-10k": Synthea COVID-19 data (10k patients)
\item "synthea-covid19-200k": Synthea COVID-19 data (200k patients)
\item "synthea-heart-10k": Synthea heart disease data
\item "synpuf-1k": CMS Synthetic Public Use Files (1k patients)
See Details for complete list.
}}

\item{cdm_version}{Character string specifying OMOP CDM version.
Must be "5.3" or "5.4". Default is "5.3".}

\item{verbose}{Logical indicating whether to print progress messages.
Default is TRUE.}
}
\value{
A list containing:
\item{cdm}{CDM reference object from CDMConnector}
\item{connection}{DBI connection to the database}
}
\description{
Creates an in-memory OMOP Common Data Model (CDM) database using Eunomia
synthetic datasets. Eunomia provides several pre-built OMOP CDM datasets
including Synthea-generated data and CMS SynPUF data.
}
\details{
Available Eunomia datasets include:
\itemize{
\item GiBleed - Default small dataset for testing
\item synthea-allergies-10k - Synthea allergies data
\item synthea-anemia-10k - Synthea anemia data
\item synthea-breast_cancer-10k - Synthea breast cancer data
\item synthea-contraceptives-10k - Synthea contraceptives data
\item synthea-covid19-10k - Synthea COVID-19 data (10k patients)
\item synthea-covid19-200k - Synthea COVID-19 data (200k patients)
\item synthea-dermatitis-10k - Synthea dermatitis data
\item synthea-heart-10k - Synthea heart disease data
\item synthea-hiv-10k - Synthea HIV data
\item synthea-lung_cancer-10k - Synthea lung cancer data
\item synthea-medications-10k - Synthea medications data
\item synthea-metabolic_syndrome-10k - Synthea metabolic syndrome data
\item synthea-opioid_addiction-10k - Synthea opioid addiction data
\item synthea-rheumatoid_arthritis-10k - Synthea rheumatoid arthritis data
\item synthea-snf-10k - Synthea skilled nursing facility data
\item synthea-surgery-10k - Synthea surgery data
\item synthea-total_joint_replacement-10k - Synthea joint replacement data
\item synpuf-1k - CMS Synthetic Public Use Files
}

The function automatically sets up the EUNOMIA_DATA_FOLDER environment
variable if not already configured. Data will be downloaded on first use.
}
\examples{
\dontrun{
# Setup with default GiBleed dataset
cdm_setup <- setup_omop_cdm_eunomia()

# Explore the data
print(cdm_setup$cdm)

# Clean up
cdmDisconnect(cdm_setup$cdm)

# Setup with COVID-19 dataset
covid_setup <- setup_omop_cdm_eunomia("synthea-covid19-10k")
cdmDisconnect(covid_setup$cdm)
}

}
\seealso{
\code{\link{explore_cdm}}, \code{\link{plot_cdm_demographics}}
}
